feat(enum): Fix cast to be correct

+++ ModelSaber.Models/Enums.cs
@@ -73,6 +73,6 @@ namespace ModelSaber.Models
             => Enum.GetValues(typeof(T)).Cast<T>().Where(v => en.HasFlag(v));
 
         public static T GetFlagFromList<T>(this IEnumerable<T> en) where T : Enum
-            => (T)(object)en.Aggregate(0, (current, v) => current | (int)(object)v);
+            => (T)(object)en.Aggregate(0, (current, v) => current | (byte)(object)v);
     }
 }
